Add moon render to calendar
Started out trying to plot it on the calendar, then quickly noted it
would always be outside the calendar's max apogee.

Then I tried scaling it down, but it still regularly ended up on top of
the calendar dates or otherwise obnoxiously placed.

Last version is simply always present outside the calendar.